+++ app/models/exchanges/quote.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module Exchanges
+  class Quote < ApplicationRecord
+    self.table_name = :exchanges_quotes
+
+    belongs_to :exchange
+    belongs_to :securitiable, polymorphic: true
+  end
+end