+++ app/models/foundation.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
 
 class Foundation < ApplicationRecord
+  include Securitiable
 end