+++ spec/factories/exchanges/quotes.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+FactoryBot.define do
+  factory :quote, class: 'Exchanges::Quote' do
+    association :exchange
+    association :securitiable, factory: :share
+  end
+end