+++ spec/models/bond_spec.rb
@@ -6,4 +6,8 @@ describe Bond, type: :model do
 
     expect(bond).to be_valid
   end
+
+  describe 'associations' do
+    it { is_expected.to have_many(:quotes).class_name('Exchanges::Quote').dependent(:destroy) }
+  end
 end